move src to netlify folder
